Moving HandleGuard into it's own file as per code review comment.
