+++ projects/miopen/test/gtest/gtest_desc_guard.hpp
@@ -53,43 +53,6 @@ using DropoutDescGuard = DescGuard<miopenDropoutDescriptor_t,
 using RNNDescGuard =
     DescGuard<miopenRNNDescriptor_t, miopenCreateRNNDescriptor, miopenDestroyRNNDescriptor>;
 
-// RAII wrapper for miopenHandle_t. Cannot use the DescGuard template because
-// miopenCreateWithStream takes an extra hipStream_t argument. The handle is
-// default-constructed empty; call create(stream) to allocate it lazily.
-class HandleGuard
-{
-public:
-    HandleGuard() = default;
-
-    explicit HandleGuard(hipStream_t stream) { status = miopenCreateWithStream(&handle, stream); }
-
-    ~HandleGuard()
-    {
-        if(handle != nullptr)
-            miopenDestroy(handle);
-    }
-
-    void create(hipStream_t stream)
-    {
-        if(handle != nullptr)
-            miopenDestroy(handle);
-        status = miopenCreateWithStream(&handle, stream);
-    }
-
-    operator miopenHandle_t() { return handle; }
-    miopenHandle_t get() { return handle; }
-    miopenStatus_t getStatus() const { return status; }
-
-    HandleGuard(const HandleGuard&)            = delete;
-    HandleGuard& operator=(const HandleGuard&) = delete;
-    HandleGuard(HandleGuard&&)                 = delete;
-    HandleGuard& operator=(HandleGuard&&)      = delete;
-
-private:
-    miopenHandle_t handle = nullptr;
-    miopenStatus_t status = miopenStatusSuccess;
-};
-
 // Frees the internal DropoutDescriptor allocated by miopen::RNNDescriptor's
 // default and 8-arg constructors. The library has no destructor for this
 // field, so any test that creates an RNN descriptor and then calls
